Small Improvements, Added weighed histogram
ATTENTION: In this version, the files are not parallelized (so as it can
run on Windows).

bigger changes:
feedback.py:
Included possibility to give mode as strings: arpe, rpe, off or random
weighed histogram: generate same plot as feedback.py, but with weighing
of the measurement results. Not fully tested